Update layers.js
I am suggesting to change the threshold for displaying POI overlay in such a way that the points would show up at a wider zoom level (zoom 10 or 11 instead of 14).  The motivation being that when we are planning a route and searching for an hotel or a gas station we probably want to see a larger span of teritory to see what our options are.

Of course, this suggestion is made from the POV of someone in North America using motorized vehicles (which could be something conflicting with cyclists requirements).