Increase pixel width of mapillary coverage
I have been using mapillary with GPX studio quite a lot to identify areas of interest on my routes.   The default width is so thin that it is difficult to see what is covered and what not.

Note:  this proposed change is done from the perspective of someone in North America where roads and coverage might have less density than in Europe - so it might not make sense to all.

Nevertheless, I thought I would propose the changes and see.